+++ CHANGELOG.md
@@ -6,6 +6,25 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+- Create default account when no account is specified in `conjurctl account create`.
+  [cyberark/conjur#2388](https://github.com/cyberark/conjur/pull/2388)
+
+### Changed
+- Changed claims mapping variable name ('mapping-claims' => 'claim-aliases').
+  [cyberark/conjur#2382](https://github.com/cyberark/conjur/pull/2382)
+
+## [1.13.2] - 2021-10-13
+
+### Security
+- GCP Authenticator: When defining the host using the instance-name annotation,
+  you now need to define at least one additional annotation.
+  [cyberark/ONYX-9442](https://ca-il-jira.il.cyber-ark.com:8443/browse/ONYX-9442)
+- Updated nokogiri to 1.12.5 in both Gemfile.lock and docs/Gemfile.lock to resolve 
+  [CVE-2021-41098](https://github.com/advisories/GHSA-2rr5-8q37-2w7h)
+  [cyberark/conjur#2376](https://github.com/cyberark/conjur/pull/2376)
+  [cyberark/conjur#2377](https://github.com/cyberark/conjur/pull/2377)
+
 ## [1.13.1] - 2021-09-13
 
 ### Fixed
@@ -697,7 +716,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - The first tagged version.
 
-[Unreleased]: https://github.com/cyberark/conjur/compare/v1.13.1...HEAD
+[Unreleased]: https://github.com/cyberark/conjur/compare/v1.13.2...HEAD
+[1.13.2]: https://github.com/cyberark/conjur/compare/v1.13.1...v1.13.2
 [1.13.1]: https://github.com/cyberark/conjur/compare/v1.13.0...v1.13.1
 [1.13.0]: https://github.com/cyberark/conjur/compare/v1.12.0...v1.13.0
 [1.12.0]: https://github.com/cyberark/conjur/compare/v1.11.7...v1.12.0